+++ Tests/Editor/IntegrationTests/BuilderDocumentTests.cs
@@ -14,6 +14,18 @@ namespace Unity.UI.Builder.EditorTests
     {
         const string k_NewUxmlFilePath = "Assets/BuildDocumentTests__TestUI.uxml";
 
+        public override void Setup()
+        {
+            base.Setup();
+
+            // Make sure there's no modified version in memory.
+            if (!EditorApplication.isPlaying)
+            {
+                AssetDatabase.ImportAsset(k_TestNoUSSDocumentUXMLFilePath, ImportAssetOptions.ForceUpdate | ImportAssetOptions.ForceSynchronousImport);
+                AssetDatabase.ImportAsset(k_TestMultiUSSDocumentUXMLFilePath, ImportAssetOptions.ForceUpdate | ImportAssetOptions.ForceSynchronousImport);
+            }
+        }
+
         protected override IEnumerator TearDown()
         {
             ForceNewDocument();
@@ -28,6 +40,8 @@ namespace Unity.UI.Builder.EditorTests
             }
 
             yield return base.TearDown();
+            DeleteTestUXMLFile();
+            DeleteTestUSSFile();
         }
 
         void CheckNoUSSDocument()
@@ -169,6 +183,73 @@ namespace Unity.UI.Builder.EditorTests
             Assert.AreEqual(asset.visualElementAssets.Count, assetCount);
             var asset2 = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(k_TestMultiUSSDocumentUXMLFilePath);
             Assert.AreEqual(asset2.visualElementAssets.Count, assetCount);
+        }
+
+        [UnityTest]
+        public IEnumerator EnsureChangesToUXMLMadeExternallyAreReloaded()
+        {
+            const string testLabelName = "externally-added-label";
+
+            CreateTestUXMLFile();
+
+            var asset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(k_TestUXMLFilePath);
+            var assetCount = asset.visualElementAssets.Count;
+            BuilderWindow.LoadDocument(asset);
+            Assert.AreEqual(BuilderWindow.document.visualTreeAsset, asset);
+
+            yield return AddTextFieldElement();
+            Assert.AreEqual(asset.visualElementAssets.Count, assetCount + 1);
+
+            // Save
+            BuilderWindow.document.SaveUnsavedChanges(k_TestUXMLFilePath, false);
+
+            var vtaCopy = BuilderWindow.document.visualTreeAsset.DeepCopy();
+            var newElement = new VisualElementAsset(typeof(Label).ToString());
+            newElement.AddProperty("name", testLabelName);
+            vtaCopy.AddElement(vtaCopy.GetRootUXMLElement(), newElement);
+            var vtaCopyUXML = vtaCopy.GenerateUXML(k_TestUXMLFilePath, true);
+            File.WriteAllText(k_TestUXMLFilePath, vtaCopyUXML);
+            AssetDatabase.ImportAsset(k_TestUXMLFilePath, ImportAssetOptions.ForceUpdate);
+
+            yield return UIETestHelpers.Pause(1);
+
+            // Make sure the UI Builder reloaded.
+            var label = BuilderWindow.documentRootElement.Q<Label>(testLabelName);
+            Assert.NotNull(label);
+        }
+
+        [UnityTest]
+        public IEnumerator EnsureChangesToUSSMadeExternallyAreReloaded()
+        {
+            const string testSelector = ".externally-added-selector";
+
+            CreateTestUXMLFile();
+            CreateTestUSSFile();
+
+            var asset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(k_TestUXMLFilePath);
+            var assetCount = asset.visualElementAssets.Count;
+            BuilderWindow.LoadDocument(asset);
+            Assert.AreEqual(BuilderWindow.document.visualTreeAsset, asset);
+
+            yield return CodeOnlyAddUSSToDocument(k_TestUSSFilePath);
+            Assert.NotNull(BuilderWindow.document.activeStyleSheet);
+
+            // Save
+            BuilderWindow.document.SaveUnsavedChanges(k_TestUXMLFilePath, false);
+
+            var styleSheetCopy = BuilderWindow.document.activeStyleSheet.DeepCopy();
+            styleSheetCopy.AddSelector(testSelector);
+            var styleSheetCopyUSS = styleSheetCopy.GenerateUSS();
+            File.WriteAllText(k_TestUSSFilePath, styleSheetCopyUSS);
+            AssetDatabase.ImportAsset(k_TestUSSFilePath, ImportAssetOptions.ForceUpdate);
+
+            yield return UIETestHelpers.Pause(1);
+
+            // Make sure the UI Builder reloaded.
+            var activeStyleSheet = BuilderWindow.document.activeStyleSheet;
+            var complexSelector = activeStyleSheet.complexSelectors.First();
+            Assert.NotNull(complexSelector);
+            Assert.AreEqual(StyleSheetToUss.ToUssSelector(complexSelector), testSelector);
         }
 
         [UnityTest]